[TASK] allows modular usage with dynamic imports
