+++ README.md
@@ -57,6 +57,17 @@ include it anywhere in your HTML.
 <script src="node_modules/configurable-date-input-polyfill/dist/configurable-date-input-polyfill.dist.js"></script>
 ```
 
+* **ES modules:**
+```js
+import supportsDateInput from 'configurable-date-input-polyfill/src';
+
+if (!supportsDateInput()) {
+  const { default: addPickerToInputs } = await import('configurable-date-input-polyfill/src/add-picker');
+
+  addPickerToInputs([yourInputElement]);
+}
+```
+
 
 * This package supports **AMD**.
